datasource/vmware: allow building on arm systems
When built on ARM, the datasource will permanently report that it is offline.
